Refactor move options class
